requirements.txt
gradio
python-dotenv
bs4
opencv-python

# for llm
openai
text_generation
huggingface_hub
langchain

# for barcode
pandas
Pyzbar

